feat(proxy): add connection management and topology visualization
Introduce a new Connections screen for viewing and managing active/closed
network connections, featuring detailed connection info and the ability to
close specific or all sessions. Implement a TopologyChart (Sankey diagram)
on the home screen to visualize network traffic flows.

Enhance profile management with auto-update functionality, allowing users
to set update intervals for remote subscriptions. Update the proxy UI to
support forced node selection (pinning) and global timeout configuration.

Improve application reliability by implementing an automatic restart service
for unexpected proxy termination and a comprehensive crash logging system.
Refactor theme implementation, migrate to the latest Miuix KMP components,
and update the core kernel to include subscription traffic information.

+++ app/src/main/kotlin/com/github/yumelira/yumebox/presentation/component/BottomBar.kt
@@ -1,30 +1,10 @@
-/*
- * This file is part of YumeBox.
- *
- * YumeBox is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program. If not, see <https://www.gnu.org/licenses/>.
- *
- * Copyright (c)  YumeLira 2025.
- *
- */
-
 package com.github.yumelira.yumebox.presentation.component
 
 import androidx.compose.animation.*
 import androidx.compose.animation.core.tween
 import androidx.compose.foundation.pager.PagerState
-import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.Composable
 import androidx.compose.runtime.compositionLocalOf
 import androidx.compose.runtime.derivedStateOf
 import androidx.compose.runtime.getValue
@@ -36,20 +16,21 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.unit.dp
 import com.github.yumelira.yumebox.presentation.icon.Yume
 import com.github.yumelira.yumebox.presentation.icon.yume.`Arrow-down-up`
+import com.github.yumelira.yumebox.presentation.icon.yume.`Package-check`
 import com.github.yumelira.yumebox.presentation.icon.yume.Bolt
 import com.github.yumelira.yumebox.presentation.icon.yume.House
-import com.github.yumelira.yumebox.presentation.icon.yume.`Package-check`
 import com.github.yumelira.yumebox.presentation.theme.AnimationSpecs
 import com.github.yumelira.yumebox.presentation.viewmodel.AppSettingsViewModel
 import com.ramcosta.composedestinations.navigation.DestinationsNavigator
+import dev.chrisbanes.haze.hazeEffect
 import dev.chrisbanes.haze.HazeState
 import dev.chrisbanes.haze.HazeStyle
-import dev.chrisbanes.haze.hazeEffect
 import dev.oom_wg.purejoy.mlang.MLang
 import org.koin.androidx.compose.koinViewModel
 import top.yukonga.miuix.kmp.basic.FloatingNavigationBar
 import top.yukonga.miuix.kmp.basic.NavigationBar
 import top.yukonga.miuix.kmp.basic.NavigationItem
+import top.yukonga.miuix.kmp.theme.MiuixTheme
 
 val LocalPagerState = compositionLocalOf<PagerState> { error("LocalPagerState is not provided") }
 val LocalHandlePageChange = compositionLocalOf<(Int) -> Unit> { error("LocalHandlePageChange is not provided") }
@@ -87,7 +68,7 @@ fun BottomBar(
         enter = fadeIn(
             animationSpec = tween(200, easing = AnimationSpecs.EnterEasing),
         ) + slideInVertically(
-            initialOffsetY = { (it / 6).toInt() },
+            initialOffsetY = { it / 6 },
             animationSpec = tween(240, easing = AnimationSpecs.EmphasizedDecelerate),
         ) + scaleIn(
             initialScale = 0.97f,
@@ -96,7 +77,7 @@ fun BottomBar(
         exit = fadeOut(
             animationSpec = tween(160, easing = AnimationSpecs.ExitEasing),
         ) + slideOutVertically(
-            targetOffsetY = { (it / 8).toInt() },
+            targetOffsetY = { it / 8 },
             animationSpec = tween(180, easing = AnimationSpecs.EmphasizedAccelerate),
         ) + scaleOut(
             targetScale = 0.98f,
